requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.12.0
psycopg2-binary>=2.9.9
streamlit>=1.31.0
tiktoken>=0.6.0
